packages/abx-spec-config/pyproject.toml
[project]
name = "abx-spec-config"
version = "0.0.1"
dependencies = [
    "abx>=0.1.0",
    "python-benedict>=0.34.0",
    "pydantic>=2.9.2",
    "pydantic-settings>=2.6.0",
    "rich>=13.9.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.entry-points.abx]
abx_spec_config = "abx_spec_config"
